+++ include/source_output_effects.hpp
@@ -18,6 +18,8 @@ class SourceOutputEffects : public PipelineBase {
     SourceOutputEffects(const std::shared_ptr<PulseManager>& pulse_manager);
     virtual ~SourceOutputEffects();
 
+    std::string log_tag;
+
     std::shared_ptr<PulseManager> pm;
 
     std::vector<std::string> plugins_order, plugins_order_old;
@@ -47,8 +49,6 @@ class SourceOutputEffects : public PipelineBase {
     sigc::signal<void, std::array<double, 2>> webrtc_output_level;
 
    private:
-    std::string log_tag = "soe: ";
-
     GSettings* soe_settings;
 
     void add_plugins_to_pipeline();